[fix] annotation 교체로 인한 코드 변경
