+++ src/main/java/com/friedNote/friedNote_backend/domain/user/presentation/UserController.java
@@ -2,7 +2,7 @@ package com.friedNote.friedNote_backend.domain.user.presentation;
 
 import com.friedNote.friedNote_backend.domain.user.application.dto.request.UserRequest;
 import com.friedNote.friedNote_backend.domain.user.application.service.UserCreateUseCase;
-import com.friedNote.friedNote_backend.domain.user.application.service.UserUpdateService;
+import com.friedNote.friedNote_backend.domain.user.application.service.UserUpdateUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class UserController {
-    private final UserUpdateService userUpdateService;
+    private final UserUpdateUseCase userUpdateUseCase;
 
     private final UserCreateUseCase userCreateUseCase;
 
@@ -22,6 +22,6 @@ public class UserController {
 
     @PostMapping("/user/update")
     public void updateUser(@RequestBody UserRequest.UserUpdateRequest userUpdateRequest) {
-        userUpdateService.userUpdate(userUpdateRequest);
+        userUpdateUseCase.userUpdate(userUpdateRequest);
     }
 }
